+++ src/main/java/step3/Application.java
@@ -1,6 +1,6 @@
 package step3;
 
-import step3.domain.Cars;
+import step3.domain.car.Cars;
 import step3.service.RacingManager;
 import step3.view.InputView;
 import step3.view.OutputView;
