+++ src/test/java/step3/domain/car/CarTest.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package step3.domain.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
